Internal

PowerFlows._first_choice_gen_id — Method

Try to make an informative one or two character name for the generator

"generator-1234-AB" -> "AB"

"123CT7" -> "7"

PowerFlows._get_load_data — Method

Obtain total load on bus b

PowerFlows._update_branch_flow! — Method

Updates the flow on the branches

PowerFlows.create_gen_ids — Method

Given a vector of Generators, create a corresponding vector of unique-per-bus PSS/E-compatible generator IDs

WRITTEN TO SPEC: PSS/E 33.3 POM page 5-14, PSS/E 34.1 DataFormat page 1-15

PowerFlows.flow_func — Method

Calculates the From - To complex power flow using external data of voltages of branch of type Line

PowerFlows.flow_func — Method

Calculates the From - To complex power flow using external data of voltages of branch of type TapTransformer

PowerFlows.flow_func — Method

Calculates the From - To complex power flow using external data of voltages of branch of type Transformer2W

PowerFlows.flow_val — Method

Calculates the From - To complex power flow (Flow injected at the bus) of branch of type Line

PowerFlows.flow_val — Method

Calculates the From - To complex power flow (Flow injected at the bus) of branch of type Line

PowerFlows.flow_val — Method

Calculates the From - To complex power flow (Flow injected at the bus) of branch of type TapTransformer

PowerFlows.flow_val — Method

Calculates the From - To complex power flow (Flow injected at the bus) of branch of type Transformer2W

PowerFlows.make_bus_name_formatter_from_metadata — Method

Given a metadata dictionary parsed from a raw_metadata_log.json, yields a function that can be passed as the bus_name_formatter kwarg to the System constructor to properly restore Sienna bus names: System("filename.raw"; bus_name_formatter = PF.make_bus_name_formatter_from_metadata(md))

PowerFlows.make_gen_name_formatter_from_metadata — Method

Given a metadata dictionary parsed from a raw_metadata_log.json, yields a function that can be passed as the gen_name_formatter kwarg to the System constructor to properly restore Sienna generator names: System("filename.raw"; gen_name_formatter = PF.make_gen_name_formatter_from_metadata(md))

PowerFlows.solve_powerflow! — Method

Evaluates the power flows on each system's branch and updates the PowerFlowData structure.

Arguments:

data::PTDFPowerFlowData: PTDFPowerFlowData structure containig all the information related to the system's power flow.

PowerFlows.solve_powerflow! — Method

Evaluates the power flows on each system's branch and updates the PowerFlowData structure.

Arguments:

data::vPTDFPowerFlowData: vPTDFPowerFlowData structure containig all the information related to the system's power flow.

PowerFlows.update_system! — Method

Modify the values in the given System to correspond to the given PowerFlowData such that if a new PowerFlowData is constructed from the resulting system it is the same as data. See also write_powerflow_solution!.

PowerFlows.write_powerflow_solution! — Method

Updates system voltages and powers with power flow results