PowerFlows

PowerFlows.run_powerflow! — Method

Solves a the power flow into the system and writes the solution into the relevant structs. Updates generators active and reactive power setpoints and branches active and reactive power flows (calculated in the From - To direction) (see flow_val)

Supports solving using Finite Differences Method (instead of using analytic Jacobian) by setting finite_diff = true. Supports passing NLsolve kwargs in the args. By default shows the solver trace.

Arguments available for nlsolve:

method : See NLSolve.jl documentation for available solvers

xtol: norm difference in x between two successive iterates under which convergence is declared. Default: 0.0.

ftol: infinite norm of residuals under which convergence is declared. Default: 1e-8.

iterations: maximum number of iterations. Default: 1_000.

store_trace: should a trace of the optimization algorithm's state be stored? Default: false.

show_trace: should a trace of the optimization algorithm's state be shown on STDOUT? Default: false.

extended_trace: should additifonal algorithm internals be added to the state trace? Default: false.

Examples

run_powerflow!(sys)
# Passing NLsolve arguments
run_powerflow!(sys, method=:newton)
# Using Finite Differences
run_powerflow!(sys, finite_diff=true)

PowerFlows.run_powerflow — Method

Similar to run_powerflow!(sys) but does not update the system struct with results. Returns the results in a dictionary of dataframes.

Examples

res = run_powerflow(sys)
# Passing NLsolve arguments
res = run_powerflow(sys, method=:newton)
# Using Finite Differences
res = run_powerflow(sys, finite_diff=true)