Internal

PowerFlows._first_choice_gen_id — Method

Try to make an informative one or two character name for the load/generator/etc.

"generator-1234-AB" -> "AB"

"123CT7" -> "7"

"load1234" -> "34"

PowerFlows._get_load_data — Method

Obtain total load on bus b

PowerFlows._psse_bus_names — Method

Given a vector of Sienna bus names, create a dictionary from Sienna bus name to PSS/E-compatible bus name. Guarantees determinism and minimal changes.

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Bus Data

PowerFlows._psse_bus_numbers — Method

Given a vector of Sienna bus numbers, create a dictionary from Sienna bus number to PSS/E-compatible bus number. Assumes that the Sienna bus numbers are positive and unique. Guarantees determinism: if the input contains the same bus numbers in the same order, the output will. Guarantees minimal changes: that if an existing bus number is compliant, it will not be changed.

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Bus Data

PowerFlows._psse_container_numbers — Method

Validate that the Sienna area/zone names parse as PSS/E-compatible area/zone numbers, output a mapping

PowerFlows._psse_transformer_names — Method

Given a vector of Sienna transformer names, create a dictionary from Sienna transformer name to PSS/E-compatible transformer name. Guarantees determinism and minimal changes.

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Transformer Data

PowerFlows._update_branch_flow! — Method

Updates the flow on the branches

PowerFlows._write_bus_data — Method

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Bus Data. Sienna voltage limits treated as PSS/E normal voltage limits; PSSE emergency voltage limits left as default.

PowerFlows._write_case_identification_data — Method

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Case Identification Data

PowerFlows._write_fixed_bus_shunt_data — Method

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Fixed Bus Shunt Data

PowerFlows._write_generator_data — Method

If the flag sources_as_generators is set, export PSY.Source instances as PSS/E generators in addition to PSY.Generators

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Fixed Bus Shunt Data

PowerFlows._write_load_data — Method

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Load Data

PowerFlows._write_non_transformer_branch_data — Method

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Fixed Bus Shunt Data

PowerFlows._write_transformer_data — Method

Currently only supports two-winding transformers

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Fixed Bus Shunt Data

PowerFlows._write_zone_data — Method

Assumes that the Sienna zone names are already PSS/E compatible

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Zone Data

PowerFlows.check_33 — Method

Throw a NotImplementedError if the psse_version is not :v33

PowerFlows.create_component_ids — Method

Given a vector of component names and a corresponding vector of container IDs (e.g., bus numbers), create unique-per-container PSS/E-compatible IDs, output a dictionary from (container ID, component name) to PSS/E-compatible component ID. The "singlesto1" flag detects components that are the only one on their bus and gives them the name "1".

PowerFlows.flow_func — Method

Calculates the From - To complex power flow using external data of voltages of branch of type Line

PowerFlows.flow_func — Method

Calculates the From - To complex power flow using external data of voltages of branch of type TapTransformer

PowerFlows.flow_func — Method

Calculates the From - To complex power flow using external data of voltages of branch of type Transformer2W

PowerFlows.flow_val — Method

Calculates the From - To complex power flow (Flow injected at the bus) of branch of type Line

PowerFlows.flow_val — Method

Calculates the From - To complex power flow (Flow injected at the bus) of branch of type Line

PowerFlows.flow_val — Method

Calculates the From - To complex power flow (Flow injected at the bus) of branch of type TapTransformer

PowerFlows.flow_val — Method

Calculates the From - To complex power flow (Flow injected at the bus) of branch of type Transformer2W

PowerFlows.joinln — Function

join with a newline at the end, delimeter defaults to ", ". If strip_trailing_empties, trailing entries of iterator equal to the empty string are ignored.

PowerFlows.make_bus_name_formatter_from_metadata — Method

Given a metadata dictionary parsed from a raw_metadata_log.json, yields a function that can be passed as the bus_name_formatter kwarg to the System constructor to properly restore Sienna bus names: System("filename.raw"; bus_name_formatter = PF.make_bus_name_formatter_from_metadata(md))

PowerFlows.make_gen_name_formatter_from_metadata — Method

Given a metadata dictionary parsed from a raw_metadata_log.json, yields a function that can be passed as the gen_name_formatter kwarg to the System constructor to properly restore Sienna generator names: System("filename.raw"; gen_name_formatter = PF.make_gen_name_formatter_from_metadata(md))

PowerFlows.solve_powerflow! — Method

Evaluates the power flows on each system's branch and updates the PowerFlowData structure.

Arguments:

data::PTDFPowerFlowData: PTDFPowerFlowData structure containig all the information related to the system's power flow.

PowerFlows.solve_powerflow! — Method

Evaluates the power flows on each system's branch and updates the PowerFlowData structure.

Arguments:

data::vPTDFPowerFlowData: vPTDFPowerFlowData structure containig all the information related to the system's power flow.

PowerFlows.update_system! — Method

Modify the values in the given System to correspond to the given PowerFlowData such that if a new PowerFlowData is constructed from the resulting system it is the same as data. See also write_powerflow_solution!.

PowerFlows.with_units — Method

A context manager similar to Logging.with_logger that sets the system's units to the given value, executes the function, then sets them back. Suppresses logging below Warn from internal calls to set_units_base_system!.

PowerFlows.write_powerflow_solution! — Method

Updates system voltages and powers with power flow results