Internal

PowerFlows.SystemPowerFlowContainer — Type

A PowerFlowContainer that represents its data as a PSY.System

PowerFlows._first_choice_gen_id — Method

Try to make an informative one or two character name for the load/generator/etc.

"generator-1234-AB" -> "AB"

"123CT7" -> "7"

"load1234" -> "34"

PowerFlows._get_load_data — Method

Obtain total load on bus b

PowerFlows._map_psse_container_names — Method

Validate that the Sienna area/zone names parse as PSS/E-compatible area/zone numbers, output a mapping

PowerFlows._psse_bus_names — Method

Given a vector of Sienna bus names, create a dictionary from Sienna bus name to PSS/E-compatible bus name. Guarantees determinism and minimal changes.

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Bus Data

PowerFlows._psse_bus_numbers — Method

Given a vector of Sienna bus numbers, create a dictionary from Sienna bus number to PSS/E-compatible bus number. Assumes that the Sienna bus numbers are positive and unique. Guarantees determinism: if the input contains the same bus numbers in the same order, the output will. Guarantees minimal changes: that if an existing bus number is compliant, it will not be changed.

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Bus Data

PowerFlows._psse_transformer_names — Method

Given a vector of Sienna transformer names, create a dictionary from Sienna transformer name to PSS/E-compatible transformer name. Guarantees determinism and minimal changes.

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Transformer Data

PowerFlows._update_branch_flow! — Method

Updates the flow on the branches

PowerFlows.better_float_to_buf — Method

Temporary, very specialized proof of concept patch for https://github.com/JuliaLang/julia/issues/55835

PowerFlows.check_33 — Method

Throw a NotImplementedError if the psse_version is not :v33

PowerFlows.convert_empty — Method

If val is empty, returns T(); if not, asserts that val isa T and returns val. Has nice type checker semantics.

Examples

convert_empty(Vector{String}, [])  # -> String[]
convert_empty(Vector{String}, ["a"])  # -> ["a"]
convert_empty(Vector{String}, [2])  # -> TypeError: in typeassert, expected Vector{String}, got a value of type Vector{Int64}
Base.return_types(Base.Fix1(convert_empty, Vector{String}))  # -> [Vector{String}]

PowerFlows.create_component_ids — Method

Given a vector of component names and a corresponding vector of container IDs (e.g., bus numbers), create unique-per-container PSS/E-compatible IDs, output a dictionary from (container ID, component name) to PSS/E-compatible component ID. The "singlesto1" flag detects components that are the only one on their bus and gives them the name "1".

PowerFlows.deserialize_reverse_component_ids — Method

Convert a sbusnsname => pname dictionary to a (pbusn, pname) => s_name dictionary

PowerFlows.fix_load_zone_names! — Method

Rename all the LoadZones in the system according to the Load_Zone_Name_Mapping in the metadata

PowerFlows.fix_nans! — Method

Iterate over all the Generators in the system and, if any active_power or reactive_power fields are NaN, make them 0.0

PowerFlows.flow_func — Method

Calculates the From - To complex power flow using external data of voltages of branch of type Line

PowerFlows.flow_func — Method

Calculates the From - To complex power flow using external data of voltages of branch of type TapTransformer

PowerFlows.flow_func — Method

Calculates the From - To complex power flow using external data of voltages of branch of type Transformer2W

PowerFlows.flow_val — Method

Calculates the From - To complex power flow (Flow injected at the bus) of branch of type Line

PowerFlows.flow_val — Method

Calculates the From - To complex power flow (Flow injected at the bus) of branch of type Line

PowerFlows.flow_val — Method

Calculates the From - To complex power flow (Flow injected at the bus) of branch of type TapTransformer

PowerFlows.flow_val — Method

Calculates the From - To complex power flow (Flow injected at the bus) of branch of type Transformer2W

PowerFlows.get_active_power_limits_for_power_flow — Method

Return the active power limits that should be used in power flow calculations and PSS/E exports. Redirects to PSY.get_active_power_limits in all but special cases.

PowerFlows.get_reactive_power_limits_for_power_flow — Method

Return the reactive power limits that should be used in power flow calculations and PSS/E exports. Redirects to PSY.get_reactive_power_limits in all but special cases.

PowerFlows.make_power_flow_container — Function

Create an appropriate PowerFlowContainer for the given PowerFlowEvaluationModel and initialize it from the given PSY.System.

Arguments:

pfem::PowerFlowEvaluationModel: power flow model to construct a container for (e.g., DCPowerFlow())

sys::PSY.System: the system from which to initialize the power flow container

time_steps::Int: number of time periods to consider (default is 1)

timestep_names::Vector{String}: names of the time periods defines by the argument "time_steps". Default value is String[].

check_connectivity::Bool: Perform connectivity check on the network matrix. Default value is true.

PowerFlows.name_formatter_from_component_ids — Method

Use PSS/E exporter metadata to build a function that maps component names back to their original Sienna values.

PowerFlows.reset_caches — Method

Force all cached information (serialized metadata, component lists, etc.) to be regenerated

PowerFlows.serialize_component_ids — Method

Take the output of create_component_ids and make it more suitable for JSON serialization

PowerFlows.supports_multi_period — Method

Trait signifying whether the PowerFlowContainer can represent multi-period data. Must be implemented for all concrete subtypes.

PowerFlows.update_system! — Method

Modify the values in the given System to correspond to the given PowerFlowData such that if a new PowerFlowData is constructed from the resulting system it is the same as data. See also write_powerflow_solution!. NOTE that this assumes that data was initialized from sys and then solved with no further modifications.

PowerFlows.write_powerflow_solution! — Method

Updates system voltages and powers with power flow results

PowerFlows.write_to_buffers! — Method

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Bus Data. Sienna voltage limits treated as PSS/E normal voltage limits; PSSE emergency voltage limits left as default.

PowerFlows.write_to_buffers! — Method

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Case Identification Data

PowerFlows.write_to_buffers! — Method

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Fixed Bus Shunt Data

PowerFlows.write_to_buffers! — Method

If the exportsettings flag `sourcesas_generatorsis set, exportPSY.Sourceinstances as PSS/E generators in addition toPSY.Generator`s

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Generator Data

PowerFlows.write_to_buffers! — Method

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Load Data

PowerFlows.write_to_buffers! — Method

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Non-Transformer Branch Data

PowerFlows.write_to_buffers! — Method

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Q Record

PowerFlows.write_to_buffers! — Method

Currently only supports two-winding transformers

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Transformer Data

PowerFlows.write_to_buffers! — Method

Assumes that the Sienna zone names are already PSS/E compatible

WRITTEN TO SPEC: PSS/E 33.3 POM 5.2.1 Zone Data