PowerFlows

PowerFlows.solve_powerflow! — Method

Evaluates the power flows on each system's branch and updates the PowerFlowData structure.

Arguments:

data::PTDFPowerFlowData: PTDFPowerFlowData structure containig all the information related to the system's power flow.

PowerFlows.solve_powerflow! — Method

Evaluates the power flows on each system's branch and updates the PowerFlowData structure.

Arguments:

data::vPTDFPowerFlowData: vPTDFPowerFlowData structure containig all the information related to the system's power flow.

PowerFlows.solve_powerflow! — Method

Solves a the power flow into the system and writes the solution into the relevant structs. Updates generators active and reactive power setpoints and branches active and reactive power flows (calculated in the From - To direction) (see flow_val)

Supports solving using Finite Differences Method (instead of using analytic Jacobian) by setting finite_diff = true. Supports passing NLsolve kwargs in the args. By default shows the solver trace.

Arguments available for nlsolve:

method : See NLSolve.jl documentation for available solvers

xtol: norm difference in x between two successive iterates under which convergence is declared. Default: 0.0.

ftol: infinite norm of residuals under which convergence is declared. Default: 1e-8.

iterations: maximum number of iterations. Default: 1_000.

store_trace: should a trace of the optimization algorithm's state be stored? Default: false.

show_trace: should a trace of the optimization algorithm's state be shown on STDOUT? Default: false.

extended_trace: should additifonal algorithm internals be added to the state trace? Default: false.

Examples

solve_powerflow!(sys)
# Passing NLsolve arguments
solve_powerflow!(sys, method=:newton)
# Using Finite Differences
solve_powerflow!(sys, finite_diff=true)

PowerFlows.solve_powerflow — Method

Similar to solve_powerflow!(sys) but does not update the system struct with results. Returns the results in a dictionary of dataframes.

Examples

res = solve_powerflow(sys)
# Passing NLsolve arguments
res = solve_powerflow(sys, method=:newton)
# Using Finite Differences
res = solve_powerflow(sys, finite_diff=true)

PowerFlows.solve_powerflow — Method

Evaluates the power flows on each system's branch by means of the ABA and BA matrices. Updates the PowerFlowData structure and returns a dictionary containing a DataFrame for the single timestep considered. The DataFrame containts the flows and angles related to the information stored in the PSY.System considered as input.

Arguments:

::DCPowerFlow: use DCPowerFlow() to evaluate the power flows according to the method based on the ABA and BA matrices

sys::PSY.System: container gathering the system data used for the evaluation of flows and angles.

PowerFlows.solve_powerflow — Method

Evaluates the power flows on each system's branch by means of the PTDF matrix. Updates the PowerFlowData structure and returns a dictionary containing a DataFrame for the single timestep considered. The DataFrame containts the flows and angles related to the information stored in the PSY.System considered as input.

Arguments:

::PTDFDCPowerFlow: use PTDFDCPowerFlow() to evaluate the power flows according to the method based on the PTDF matrix

sys::PSY.System: container gathering the system data used for the evaluation of flows and angles.

PowerFlows.solve_powerflow — Method

Evaluates the power flows on each system's branch by means of Virtual PTDF matrices. Updates the PowerFlowData structure "data" and returns a dictionary containing a number of DataFrames equal to the numeber of timestep considered in "data". Each DataFrame containts the flows and angles.

Arguments:

data::PTDFPowerFlowData: PowerFlowData structure containing the system data per each timestep considered, as well as the Virtual PTDF matrix.

sys::PSY.System: container gathering the system data.

PowerFlows.solve_powerflow — Method

Evaluates the power flows on each system's branch by means of the ABA and BA matrices. Updates the PowerFlowData structure "data" and returns a dictionary containing a number of DataFrames equal to the numeber of timestep considered in "data". Each DataFrame containts the flows and angles.

Arguments:

data::ABAPowerFlowData: PowerFlowData structure containing the system's data per each timestep considered, as well as the ABA and BA matrices.

sys::PSY.System: container gathering the system data.

PowerFlows.solve_powerflow — Method

Evaluates the power flows on each system's branch by means of the PTDF matrix. Updates the PowerFlowData structure "data" and returns a dictionary containing a number of DataFrames equal to the numeber of timestep considered in "data". Each DataFrame containts the flows and angles.

Arguments:

data::PTDFPowerFlowData: PowerFlowData structure containing the system's data per each timestep considered, as well as the PTDF matrix.

sys::PSY.System: container gathering the system data.

PowerFlows.solve_powerflow — Method

Evaluates the power flows on each system's branch by means of the Virtual PTDF matrix. Updates the PowerFlowData structure "data" and returns a dictionary containing a number of DataFrames equal to the numeber of timestep considered in "data". The DataFrame containts the flows and angles related to the information stored in the PSY.System considered as input.

Arguments:

::vPTDFDCPowerFlow: use vPTDFDCPowerFlow() to evaluate the power flows according to the method based on the Virtual PTDF matrix

sys::PSY.System: container gathering the system data used for the evaluation of flows and angles.

PowerFlows.write_results — Method

Return power flow results in dictionary of dataframes.