PowerFlows

Powerflow Evalution Models and AC Solvers

PowerFlows.ACPowerFlow — Type

A struct for evaluating power flow solutions in AC systems.

This struct is parameterized by the type of AC power flow solver to use, which must be a subtype of ACPowerFlowSolverType. It also contains a few fields that control whether to compute certain additional data, like loss factors: see the constructor for details.

PowerFlows.ACPowerFlow — Method

ACPowerFlow{ACSolver}(
    check_reactive_power_limits::Bool = false,
    exporter::Union{Nothing, PowerFlowEvaluationModel} = nothing,
    calculate_loss_factors::Bool = false,
    generator_slack_participation_factors::Union{
        Nothing,
        Dict{Tuple{DataType, String}, Float64},
        Vector{Dict{Tuple{DataType, String}, Float64}},
    } = nothing,
) where {ACSolver <: ACPowerFlowSolverType}

An evaluation model for a standard AC powerflow with the specified solver type.

Arguments

ACSolver: The type of AC power flow solver to use, which must be a subtype of ACPowerFlowSolverType. Default is NewtonRaphsonACPowerFlow.

check_reactive_power_limits::Bool: Whether to check reactive power limits during the power flow solution. Default is false.

exporter::Union{Nothing, PowerFlowEvaluationModel}: An optional exporter for the power flow results. If not nothing, it should be a PSSEExportPowerFlow.

calculate_loss_factors::Bool: Whether to calculate loss factors during the power flow solution. Default is false.

generator_slack_participation_factors::Union{Nothing, Dict{Tuple{DataType, String}, Float64}, Vector{Dict{Tuple{DataType, String}, Float64}}}: An optional parameter that specifies the participation factors for generator slack in the power flow solution. If nothing, all slack is picked up by the reference bus. If a Dict, it should map (component_type, component_name) tuples to participation factors. If a Vector, it should contain multiple such dictionaries, allowing for different participation factors for different time steps.

PowerFlows.ACPowerFlowSolverType — Type

An abstract supertype for all iterative methods. Subtypes: NewtonRaphsonACPowerFlow, TrustRegionACPowerFlow, LevenbergMarquardtACPowerFlow, and RobustHomotopyPowerFlow.

PowerFlows.DCPowerFlow — Type

DCPowerFlow(
    exporter::Union{Nothing, PowerFlowEvaluationModel} = nothing,
)

An evaluation model for a standard DC powerflow.

This provides a fast approximate solution to the AC powerflow problem, by solving for the bus voltage angles under some simplifying assumptions (lossless lines, constant voltage magnitudes, etc.). For details, see Wikipedia or section 4 of the MATPOWER docs. If not nothing, the exporter should be a PSSEExportPowerFlow.

PowerFlows.LevenbergMarquardtACPowerFlow — Type

LevenbergMarquardtACPowerFlow <: ACPowerFlowSolverType

An ACPowerFlowSolverType corresponding to the Levenberg-Marquardt iterative method.

PowerFlows.NewtonRaphsonACPowerFlow — Type

NewtonRaphsonACPowerFlow <: ACPowerFlowSolverType

An ACPowerFlowSolverType corresponding to a basic Newton-Raphson iterative method. The Newton step is taken verbatim at each iteration: no line search is performed.

PowerFlows.PTDFDCPowerFlow — Type

PTDFDCPowerFlow(
    exporter::Union{Nothing, PowerFlowEvaluationModel} = nothing,
)

An evaluation model that calculates line flows using the Power Transfer Distribution Factor Matrix.

This approximates the branch flows in the power grid, under some simplifying assumptions (lossless lines, constant voltage magnitudes, etc.). See section 4 of the MATPOWER docs for details. If not nothing, the exporter should be a PSSEExportPowerFlow.

PowerFlows.RobustHomotopyPowerFlow — Type

RobustHomotopyPowerFlow <: ACPowerFlowSolverType

An ACPowerFlowSolverType corresponding to a homotopy iterative method, based on the paper "Improving the robustness of Newton-based power flow methods to cope with poor initial points".

PowerFlows.TrustRegionACPowerFlow — Type

TrustRegionACPowerFlow <: ACPowerFlowSolverType

An ACPowerFlowSolverType corresponding to the Powell dogleg iterative method.

PowerFlows.vPTDFDCPowerFlow — Type

vPTDFDCPowerFlow(
    exporter::Union{Nothing, PowerFlowEvaluationModel} = nothing,
)

An evaluation model that calculates line flows using a virtual Power Transfer Distribution Factor Matrix.

This is a replacement for the PTDFDCPowerFlow for large grids, where creating and storing the full PTDF matrix would be infeasible or slow. See the PowerNetworkMatrices.jl docs for details. If not nothing, the exporter should be a PSSEExportPowerFlow.

PowerFlows.PSSEExportPowerFlow — Method

PSSEExportPowerFlow(psse_version::Symbol, export_dir::AbstractString; kwargs...)

An evaluation model for exporting power flow results to PSSE format.

Arguments:

psse_version::Symbol: The version of PSSE to export to. Must be among [:v33, :v35].

export_dir::AbstractString: The directory where the PSSE files will be exported.

Optional keyword arguments:

name::AbstractString: The base name for the exported files. Defaults to "export".

write_comments::Bool: Whether to write comments in the exported files. Defaults to false.

overwrite::Bool: Whether to overwrite the file if it exists already. Defaults to false.

PowerFlows.PSSEExporter — Type

Structure to perform an export from a Sienna System, plus optional updates from PowerFlowData, to the PSS/E format.

Construct this object from a System and a PSS/E version, update using update_exporter with any new data as relevant, and perform the export with write_export. Writes a <name>.raw file and a <name>_export_metadata.json file with transformations that had to be made to conform to PSS/E naming rules, which can be parsed by PowerSystems.jl to perform a round trip with the names restored.

Arguments:

base_system::PSY.System: the system to be exported. Later updates may change power flow-related values but may not fundamentally alter the system

psse_version::Symbol: the version of PSS/E to target, must be one of [:v33, :v35]

write_comments::Bool = false: whether to add the customary-but-not-in-spec-annotations after a slash on the first line and at group boundaries

name::AbstractString = "export": the base name of the export

step::Any = nothing: optional step data to append to the base export name. User is responsible for updating the step data. If the step data is nothing, it is not used; if it is a tuple or vector, it is joined with _ and concatted; else it is concatted after _.

overwrite::Bool = false: true to silently overwrite existing exports, false to throw an error if existing results are encountered

PowerFlows.get_psse_export_paths — Method

Calculate the paths of the (raw, metadata) files that would be written by a certain call to write_export

PowerFlows.update_exporter! — Method

Update the PSSEExporter with new data.

Arguments:

exporter::PSSEExporter: the exporter to update

data::PSY.PowerFlowData: the new data. Must correspond to the System with which the exporter was constructed.

PowerFlows.update_exporter! — Method

Update the PSSEExporter with new data.

Arguments:

exporter::PSSEExporter: the exporter to update

data::PSY.System: system containing the new data. Must be fundamentally the same System as the one with which the exporter was constructed, just with different values — this is the user's responsibility, we do not exhaustively verify it.

PowerFlows.write_export — Method

Perform an export from the data contained in a PSSEExporter to the PSS/E file format.

Solving Powerflows

PowerFlows.solve_powerflow! — Method

solve_powerflow!(data::PTDFPowerFlowData)

Evaluates the PTDF power flow and writes the result to the fields of the PTDFPowerFlowData structure.

This function modifies the following fields of data, setting them to the computed values:

data.bus_angles: the bus angles for each bus in the system.

data.branch_activepower_flow_from_to: the active power flow from the "from" bus to the "to" bus of each branch

data.branch_activepower_flow_to_from: the active power flow from the "to" bus to the "from" bus of each branch

Additionally, it sets data.converged to true, indicating that the power flow calculation was successful.

PowerFlows.solve_powerflow! — Method

solve_powerflow!(data::vPTDFPowerFlowData)

Evaluates the virtual PTDF power flow and writes the results to the fields of the vPTDFPowerFlowData structure.

This function modifies the following fields of data, setting them to the computed values:

data.bus_angles: the bus angles for each bus in the system.

data.branch_activepower_flow_from_to: the active power flow from the "from" bus to the "to" bus of each branch

data.branch_activepower_flow_to_from: the active power flow from the "to" bus to the "from" bus of each branch

Additionally, it sets data.converged to true, indicating that the power flow calculation was successful.

PowerFlows.solve_powerflow — Method

Evaluates the power flows on the system's branches by means of the method associated with the PowerFlowData structure data, which can be one of PTDFPowerFlowData, vPTDFPowerFlowData, or ABAPowerFlowData. Returns a dictionary of DataFrames, each containing the branch flows and bus voltages for the input PSY.System at that timestep.

Arguments:

data::Union{PTDFPowerFlowData, vPTDFPowerFlowData, ABAPowerFlowData}: PowerFlowData structure containing the system's data per each timestep considered, as well as the associated matrix for the power flow.

sys::PSY.System: container gathering the system data.

Note that data must have been created from the System sys using one of the PowerFlowData constructors.

Example

using PowerFlows, PowerSystemCaseBuilder
sys = build_system(PSITestSystems, "c_sys14")
data = PowerFlowData(PTDFDCPowerFlow(), sys, time_steps = 2)
d = solve_powerflow(data, sys)
display(d["2"]["flow_results"])

PowerFlows.solve_powerflow — Method

solve_powerflow(
    ::T,
    sys::PSY.System;
) where T <: Union{PTDFDCPowerFlow, vPTDFDCPowerFlow, DCPowerFlow}

Evaluates the provided DC power flow method T on the PowerSystems.System sys, returning a dictionary of DataFrames containing the calculated branch flows and bus angles.

Provided for convenience: this interface bypasses the need to create a PowerFlowData struct, but that's still what's happening under the hood.

Example

using PowerFlows, PowerSystemCaseBuilder
sys = build_system(PSITestSystems, "c_sys5")
d = solve_powerflow(DCPowerFlow(), sys)
display(d["1"]["flow_results"])
display(d["1"]["bus_results"])

PowerFlows.solve_powerflow! — Method

solve_powerflow!(pf::ACPowerFlow{<:ACPowerFlowSolverType}, system::PSY.System; kwargs...)

Solves the power flow in the system and writes the solution into the relevant structs. Updates active and reactive power setpoints for generators and active and reactive power flows for branches (calculated in the From - To direction and in the To - From direction).

Supports passing kwargs to the PF solver.

The bus types can be changed from PV to PQ if the reactive power limits are violated.

Arguments

pf::ACPowerFlow{<:ACPowerFlowSolverType}: The power flow solver instance.

system::PSY.System: The power system model, a PowerSystems.System struct.

kwargs...: Additional keyword arguments.

Keyword Arguments

check_connectivity::Bool: Checks if the grid is connected. Default is true.

check_reactive_power_limits::Bool: if true, the reactive power limits are enforced by changing the respective bus types from PV to PQ. Default is false.

tol: Infinite norm of residuals under which convergence is declared. Default is 1e-9.

maxIterations: Maximum number of Newton-Raphson iterations. Default is 30.

Returns

converged::Bool: Indicates whether the power flow solution converged.

The power flow results are written into the system struct.

Examples

solve_powerflow!(pf, sys)

# Passing kwargs
solve_powerflow!(pf, sys; correct_bustypes = true)

# Passing keyword arguments
solve_powerflow!(pf, sys; maxIterations=100)

PowerFlows.solve_powerflow! — Method

solve_powerflow!(data::ACPowerFlowData; pf::ACPowerFlow{<:ACPowerFlowSolverType} = ACPowerFlow(), kwargs...)

Solve the multiperiod AC power flow problem for the given power flow data.

The bus types can be changed from PV to PQ if the reactive power limits are violated.

Arguments

data::ACPowerFlowData: The power flow data containing the grid information and initial conditions.

pf::ACPowerFlow{<:ACPowerFlowSolverType}: The power flow solver type. Defaults to NewtonRaphsonACPowerFlow.

kwargs...: Additional keyword arguments.

Keyword Arguments

check_connectivity::Bool: Checks if the grid is connected. Default is true.

check_reactive_power_limits::Bool: if true, the reactive power limits are enforced by changing the respective bus types from PV to PQ. Default is false.

time_steps: Specifies the time steps to solve. Defaults to sorting and collecting the keys of data.timestep_map.

Description

This function solves the AC power flow problem for each time step specified in data. It preallocates memory for the results and iterates over the sorted time steps. For each time step, it calls the _ac_powerflow function to solve the power flow equations and updates the data object with the results. If the power flow converges, it updates the active and reactive power injections, as well as the voltage magnitudes and angles for different bus types (REF, PV, PQ). If the power flow does not converge, it sets the corresponding entries in data to NaN. Finally, it calculates the branch power flows and updates the data object.

Notes

If the grid topology changes (e.g., tap positions of transformers or in-service status of branches), the admittance matrices Yft and Ytf must be updated.

If Yft and Ytf change between time steps, the branch flow calculations must be moved inside the loop.

Examples

solve_powerflow!(data)

PowerFlows.solve_powerflow — Method

Similar to solve_powerflow!(pf, sys) but does not update the system struct with results. Returns the results in a dictionary of dataframes.

Examples

res = solve_powerflow(pf, sys)

PowerFlows.write_results — Method

write_results(
    ::ACPowerFlow{<:ACPowerFlowSolverType},
    sys::PSY.System,
    data::ACPowerFlowData,
    time_step::Int64,
) -> Dict{String, DataFrames.DataFrame}

Returns a dictionary containing the AC power flow results.

Only single-period evaluation is supported at the moment for AC Power flows. The resulting dictionary will therefore feature just one key linked to one DataFrame.

Arguments:

::ACPowerFlow: use ACPowerFlow() storing AC power flow results.

sys::PSY.System: container storing the system information.

result::Vector{Float64}: vector containing the results for one single time-period.

PowerFlows.write_results — Method

write_results(
    data::Union{PTDFPowerFlowData, vPTDFPowerFlowData, ABAPowerFlowData},
    sys::PSY.System,
)

Returns a dictionary containing the DC power flow results. Each key corresponds to the name of the considered time periods, storing a DataFrame with the powerflow results.

Arguments:

data::Union{PTDFPowerFlowData, vPTDFPowerFlowData, ABAPowerFlowData}: PowerFlowData structure containing power flows and bus angles.

sys::PSY.System: A PowerSystems.System object storing the system information.

PowerFlowData

PowerFlows.ABAPowerFlowData — Type

A type alias for a PowerFlowData struct whose type parameters are configured for the DCPowerFlow method.

PowerFlows.ACPowerFlowData — Type

A type alias for a PowerFlowData struct whose type parameters are configured for the ACPowerFlow method.

PowerFlows.PowerFlowData — Type

PowerFlowData{M <: PNM.PowerNetworkMatrix, N <: Union{PNM.PowerNetworkMatrix, Nothing}}

Structure containing all the data required for the evaluation of the power flows and angles, as well as these ones.

All fields starting with bus_ are ordered according to bus_lookup, and all fields starting with arc_ are ordered according to arc_lookup: one row per bus/arc, one column per time period. Here, buses should be understood as "buses remaining, after the network reduction." Similarly, we use "arcs" instead of "branches" to distinguish between network elements (post-reduction) and system objects (pre-reduction).

Generally, do not construct this directly. Instead, use one of the later constructors to pass in a PowerFlowEvaluationModel and a PowerSystems.System. aux\_network\_matrix and power\_network\_matrix will then be set to the appropriate matrices that are needed for computing that type of powerflow. See also ACPowerFlowData, ABAPowerFlowData, PTDFPowerFlowData, and vPTDFPowerFlowData: these are all aliases for PowerFlowData{N, M} with specific N,M, that are used for the respective type of power flow evaluations.

Fields:

bus_activepower_injection::Matrix{Float64}: matrix containing the bus active power injection.

bus_reactivepower_injection::Matrix{Float64}: matrix containing the bus reactive power injection.

bus_activepower_withdrawals::Matrix{Float64}: matrix containing the bus reactive power withdrawals.

bus_reactivepower_withdrawals::Matrix{Float64}: matrix containing the bus reactive power withdrawals.

bus_activepower_constant_current_withdrawals::Matrix{Float64}: matrix containing the bus active power constant current withdrawals.

bus_reactivepower_constant_current_withdrawals::Matrix{Float64}: matrix containing the bus reactive power constant current withdrawals.

bus_activepower_constant_impedance_withdrawals::Matrix{Float64}: matrix containing the bus active power constant impedance withdrawals.

bus_reactivepower_constant_impedance_withdrawals::Matrix{Float64}: matrix containing the bus reactive power constant impedance withdrawals.

bus_reactivepower_bounds::Matrix{Float64}: matrix containing upper and lower bounds for the reactive supply at each bus at each time period.

bus_type::Matrix{PSY.ACBusTypes}: matrix containing type of buses present in the system.

bus_magnitude::Matrix{Float64}: matrix containing the bus voltage magnitudes.

bus_angles::Matrix{Float64}: matrix containing the bus voltage angles.

arc_activepower_flow_from_to::Matrix{Float64}: matrix containing the active power flows measured at the from bus.

arc_reactivepower_flow_from_to::Matrix{Float64}: matrix containing the reactive power flows measured at the from bus.

arc_activepower_flow_to_from::Matrix{Float64}: matrix containing the active power flows measured at the to bus.

arc_reactivepower_flow_to_from::Matrix{Float64}: matrix containing the reactive power flows measured at the to bus.

timestep_map::Dict{Int, String}: dictionary mapping the number of the time periods (corresponding to the column number of the previously mentioned matrices) and their names.

power_network_matrix::M: matrix used for the evaluation of either the power flows or bus angles, depending on the method considered.

aux_network_matrix::N: matrix used for the evaluation of either the power flows or bus angles, depending on the method considered.

neighbors::Vector{Set{Int}}: Vector with the sets of adjacent buses.

PowerFlows.PowerFlowData — Method

PowerFlowData(
    pf::ACPowerFlow{<:ACPowerFlowSolverType},
    sys::PSY.System;
    time_steps::Int = 1,
    timestep_names::Vector{String} = String[],
    check_connectivity::Bool = true
) -> ACPowerFlowData{<:ACPowerFlowSolverType}

Creates the structure for an AC power flow calculation, given the System sys, the number of time periods to consider, and their names.

Calling this function will not evaluate the power flows and angles. Note that first input is of type ACPowerFlow: this version is used to solve AC powerflows, and returns an ACPowerFlowData object.

Arguments:

pf::ACPowerFlow: the settings for the AC power flow solver.

sys::PSY.System: A System object that represents the power grid under consideration.

time_steps::Int: number of time periods to consider in the PowerFlowData structure. It defines the number of columns of the matrices used to store data. Default value = 1.

timestep_names::Vector{String}: names of the time periods defined by the argument time_steps. Default value = String[].

check_connectivity::Bool: Perform connectivity check on the network matrix. Default value = true.

WARNING: functions for the evaluation of the multi-period AC PF still to be implemented.

PowerFlows.PowerFlowData — Method

PowerFlowData(
    ::DCPowerFlow,
    sys::PSY.System;
    time_steps::Int = 1,
    timestep_names::Vector{String} = String[],
    check_connectivity::Bool = true
) -> ABAPowerFlowData

Creates a PowerFlowData structure configured for a stnadard DC power flow calculation, given the System sys, the number of time periods to consider, and their names.

Calling this function will not evaluate the power flows and angles. Note that first input is of type DCPowerFlow: this version is used to solve DC powerflows, and returns an ABAPowerFlowData object.

Arguments:

::DCPowerFlow: Run a DC powerflow: internally, store the ABA matrix as power_network_matrix and the BA matrix as aux_network_matrix.

sys::PSY.System: A System object that represents the power grid under consideration.

time_steps::Int: number of time periods to consider in the PowerFlowData structure. It defines the number of columns of the matrices used to store data. Default value = 1.

timestep_names::Vector{String}: names of the time periods defined by the argument time_steps. Default value = String[].

check_connectivity::Bool: Perform connectivity check on the network matrix. Default value = true.

PowerFlows.PowerFlowData — Method

function PowerFlowData(
    ::PTDFDCPowerFlow,
    sys::PSY.System;
    time_steps::Int = 1,
    timestep_names::Vector{String} = String[]
) -> PTDFPowerFlowData

Creates a PowerFlowData structure configured for a Partial Transfer Distribution Factor Matrix DC power flow calculation, given the System sys, the number of time periods to consider, and their names.

Calling this function will not evaluate the power flows and angles. Note that first input is of type PTDFDCPowerFlow: this version is used to solve DC powerflows via the Power Transfer Distribution Factor (PTDF) matrix. This function returns an PTDFPowerFlowData object.

Arguments:

::PTDFDCPowerFlow: Run a DC powerflow with PTDF matrix: internally, store the PTDF matrix as power_network_matrix and the ABA matrix as aux_network_matrix.

sys::PSY.System: A System object that represents the power grid under consideration.

time_steps::Int: number of time periods to consider in the PowerFlowData structure. It defines the number of columns of the matrices used to store data. Default value = 1.

timestep_names::Vector{String}: names of the time periods defined by the argument time_steps. Default value = String[].

PowerFlows.PowerFlowData — Method

function PowerFlowData(
    ::vPTDFDCPowerFlow,
    sys::PSY.System;
    time_steps::Int = 1,
    timestep_names::Vector{String} = String[]
) -> vPTDFPowerFlowData

Creates a PowerFlowData structure configured for a virtual Partial Transfer Distribution Factor Matrix DC power flow calculation, given the System sys, the number of time periods to consider, and their names.

Calling this function will not evaluate the power flows and angles. Note that first input is of type vPTDFDCPowerFlow: this version is used to solve DC powerflows using a virtual Power Transfer Distribution Factor (PTDF) matrix. This function returns a vPTDFPowerFlowData object.

For internal usage: generally, do not construct this directly.

Arguments:

::PTDFDCPowerFlow: Run a virtual PTDF powerflow: internally, store the virtual PTDF matrix power_network_matrix and the ABA matrix as aux_network_matrix.

sys::PSY.System: A System object that represents the power grid under consideration.

time_steps::Int: number of time periods to consider in the PowerFlowData structure. It defines the number of columns of the matrices used to store data. Default value = 1.

timestep_names::Vector{String}: names of the time periods defined by the argument "time_steps". Default value = String[].

PowerFlows.PowerFlowData — Method

Sets the two PowerNetworkMatrix fields and a few others (timesteps, timestep_map), then creates arrays of default values (usually zeros) for the rest.